+++ automl_alex/automl_alex.py
@@ -1,8 +1,9 @@
 from sklearn.metrics import *
 from tqdm import tqdm
-import modin.pandas as pd
+import pandas as pd
 import time
 import joblib
+import automl_alex
 from .models import *
 from .cross_validation import *
 from .data_prepare import *
@@ -11,6 +12,7 @@ from .optimizer import *
 from pathlib import Path
 from .logger import *
 
+TMP_FOLDER = '.automl-alex_tmp/'
 
 ##################################### BestSingleModel ################################################
 
@@ -65,7 +67,7 @@ class ModelsReview(object):
         score_ls = []
         time_ls = []
         if models_names is None:
-            self.models_names = all_models.keys()
+            self.models_names = automl_alex.models.all_models.keys()
         else:
             self.models_names = models_names
 
@@ -78,12 +80,12 @@ class ModelsReview(object):
         for model_name in tqdm(self.models_names, disable=disable_tqdm):
             # Model
             start_time = time.time()
-            model_tmp = all_models[model_name](
+            model_tmp = automl_alex.models.all_models[model_name](
                                             gpu=self._gpu, 
                                             random_state=self._random_state,
                                             type_of_estimator=self.type_of_estimator)
             # fit
-            model_tmp = model_tmp.fit(X_train, y_train)
+            model_tmp.fit(X_train, y_train)
             # Predict
             if (self._metric.__name__ in predict_proba_metrics) and (model_tmp.is_possible_predict_proba()):
                 y_pred = model_tmp.predict_proba(X_test)
@@ -171,7 +173,7 @@ class AutoML(object):
     def fit(self,
         X, 
         y,
-        timeout=300, # optimization time in seconds
+        timeout=500, # optimization time in seconds
         auto_parameters=True,
         folds=5,
         score_folds=2,
@@ -186,16 +188,20 @@ class AutoML(object):
         # STEP 0
         start_step_0 = time.time()
         logger.info('> Start Fit Base Model')
+        if timeout < 400:
+            logger.warning("! Not enough time to find the optimal parameters. \n \
+                    Please, Increase the 'timeout' parameter for normal optimization. (min 500 sec)")
 
         self.cat_features=X.columns[(X.dtypes == 'object') | (X.dtypes == 'category')]
         X[self.cat_features] = X[self.cat_features].astype('str')
         X.fillna(0, inplace=True)
         X[self.cat_features] = X[self.cat_features].astype('category')
 
-        self.model_1 = CatBoost(
+        self.model_1 = automl_alex.CatBoost(
             type_of_estimator=self.type_of_estimator, 
             random_state=self._random_state,
             gpu=self._gpu,
+            verbose=verbose,
             )
         self.model_1 = self.model_1.fit(X, y, cat_features=self.cat_features.tolist())
         X = None
@@ -204,7 +210,7 @@ class AutoML(object):
         # STEP 1
         start_step_1 = time.time()
         logger.info('> DATA PREPROC')
-        self.de_1 = DataPrepare(
+        self.de_1 = automl_alex.DataPrepare(
             cat_encoder_names=['OneHotEncoder', 'CountEncoder'],
             #outliers_threshold=3,
             normalization=True,
@@ -215,18 +221,41 @@ class AutoML(object):
         if self.de_1.cat_features is not None:
             X = X.drop(self.de_1.cat_features, axis = 1)
 
-        logger.info('> Start Fit Models 1')
-        # Model 2
-        # self.model_2 = LinearModel(
-        #     type_of_estimator=self.type_of_estimator, 
-        #     random_state=self._random_state,
-        #     )
-        # self.model_2 = self.model_2.fit(X, y)
+        params = {
+                'metric' : self.metric,
+                'metric_round' :self._metric_round,
+                'auto_parameters':auto_parameters,
+                'folds':folds,
+                'score_folds':score_folds,
+                'opt_lvl':opt_lvl,
+                'early_stoping':early_stoping,
+                'type_of_estimator':self.type_of_estimator,
+                'random_state':self._random_state,
+                'gpu':self._gpu,
+                #'iteration_check': False,
+                }
 
+        # logger.info(50*'#')
+        # logger.info('> Start Fit Models 2')
+        # logger.info(50*'#')
+        # # Model 2
+        # self.model_2 = automl_alex.BestSingleModel(
+        #     models_names = ['LinearModel',],
+        #     feature_selection=False,
+        #     **params,
+        #     )
+
+        # history = self.model_2.opt(X,y, timeout=100, verbose=verbose)
+        # self.model_2.save(name='model_2', folder=TMP_FOLDER,)
+
+        logger.info(50*'#')
+        logger.info('> Start Fit Models 3')
+        logger.info(50*'#')
         # Model 3
-        self.model_3 = MLP(
+        self.model_3 = automl_alex.MLP(
             type_of_estimator=self.type_of_estimator, 
-            random_state=self._random_state
+            random_state=self._random_state,
+            verbose=verbose,
             )
         self.model_3 = self.model_3.fit(X, y)
 
@@ -250,12 +279,14 @@ class AutoML(object):
         X = self.de_2.fit_transform(X_source)
         #X = X.drop(self.de_2.cat_features, axis = 1)
 
-        logger.info('> Start Fit Models 2')
+        logger.info(50*'#')
+        logger.info('> Start Fit Models 4')
 
-        self.model_4 = CatBoost(
+        self.model_4 = automl_alex.CatBoost(
             type_of_estimator=self.type_of_estimator, 
             random_state=self._random_state,
             gpu=self._gpu,
+            verbose=verbose,
             )
 
         self.model_4 = self.model_4.fit(X, y)
@@ -268,58 +299,24 @@ class AutoML(object):
         start_step_3 = time.time()
 
         logger.info(50*'#')
-        logger.info('> Start Opt Model')
+        logger.info('> Start Fit Models 5')
+        logger.info(50*'#')
 
         time_to_opt = (timeout - (time.time()-start_step_0)) - 60
         time.sleep(0.1)
 
-        self.model_5 = LightGBM(
-            type_of_estimator=self.type_of_estimator, 
-            random_state=self._random_state,
-            gpu=self._gpu,
+        self.model_5 = automl_alex.BestSingleModel(
+            models_names = ['LightGBM', 'ExtraTrees'],
+            feature_selection=feature_selection,
+            **params,
             )
 
-        params = {'X':X,
-                'y':y,
-                'metric' : self.metric,
-                'combined_score_opt' : self._combined_score_opt,
-                'metric_round' :self._metric_round,
-                'auto_parameters':auto_parameters,
-                'folds':folds,
-                'score_folds':score_folds,
-                'opt_lvl':opt_lvl,
-                'early_stoping':early_stoping,
-                'feature_selection':feature_selection,
-                'timeout':time_to_opt, # optimization time in seconds
-                'iteration_check': False,
-                'verbose':verbose}
-
-        history = self.model_5.opt(**params)
-            
-        self._select_features_model_5 = list(history['columns'].iloc[0])
-        self.best_params_model_5 = history['model_param'].iloc[0]
-        folds=self.model_5._folds
-
-        self.model_5 = LightGBM(
-            type_of_estimator=self.type_of_estimator, 
-            random_state=self._random_state,
-            gpu=self._gpu,
-            )
-        self.model_5.model_param = self.best_params_model_5
-
-        self.cv_model_5 = CrossValidation(
-        estimator=self.model_5,
-        folds=folds,
-        n_repeats=1,
-        print_metric=False, 
-        random_state=self._random_state,
-        )
-        self.cv_model_5.fit(X[self._select_features_model_5], y,)
+        history = self.model_5.opt(X,y, timeout=time_to_opt, verbose=verbose)
+        self.model_5.save(name='model_5', folder=TMP_FOLDER,)
 
         total_step_4 = (time.time() - start_step_3)
 
         ####################################################
-
         logger.info(50*'#')
         logger.info('> Finish!')
 
@@ -342,6 +339,7 @@ class AutoML(object):
         X.fillna(0, inplace=True)
         X[self.cat_features] = X[self.cat_features].astype('category')
 
+        # MODEL 1
         self.predict_model_1 = self.model_1.predict_or_predict_proba(X)
 
         ####################################################
@@ -350,7 +348,11 @@ class AutoML(object):
         if self.de_1.cat_features is not None:
             X = X.drop(self.de_1.cat_features, axis = 1)
 
-        #self.predict_model_2 = self.model_2.predict_or_predict_proba(X)
+        # MODEL 2
+        # self.model_2 = self.model_2.load(name='model_2', folder=TMP_FOLDER,)
+        # self.predict_model_2 = self.model_2.predict(X)
+
+        # MODEL 3
         self.predict_model_3 = self.model_3.predict_or_predict_proba(X)
 
         ####################################################
@@ -358,11 +360,12 @@ class AutoML(object):
         X = self.de_2.transform(X_source, verbose=verbose)
         #X = X.drop(self.de_2.cat_features, axis = 1)
 
+        # MODEL 4
         self.predict_model_4 = self.model_4.predict_or_predict_proba(X)
         
-        ####################################################
-        # STEP 3
-        self.predict_model_5 = self.cv_model_5.predict_test(X[self._select_features_model_5])
+        # MODEL 5
+        self.model_5 = self.model_5.load(name='model_5', folder=TMP_FOLDER,)
+        self.predict_model_5 = self.model_5.predict(X)
         
         ####################################################
         # STEP 4
@@ -383,7 +386,8 @@ class AutoML(object):
         self.de_1.save(name='DataPrepare_1_dump', folder=dir_tmp)
         self.de_2.save(name='DataPrepare_2_dump', folder=dir_tmp)
         joblib.dump(self, dir_tmp+'AutoML'+'.pkl')
-        self.cv_model_5.save(name='cv_model_5', folder=dir_tmp,)
+        #self.model_2.save(name='model_2', folder=dir_tmp,)
+        self.model_5.save(name='model_5', folder=dir_tmp,)
         shutil.make_archive(folder+name, 'zip', dir_tmp)
         shutil.rmtree(dir_tmp)
         logger.info('Save AutoML')
@@ -399,7 +403,8 @@ class AutoML(object):
         model.de_1 = model.de_1.load('DataPrepare_1_dump', folder=dir_tmp)
         model.de_2 = DataPrepare()
         model.de_2 = model.de_2.load('DataPrepare_2_dump', folder=dir_tmp)
-        model.cv_model_5 = model.cv_model_5.load(name='cv_model_5', folder=dir_tmp,)
+        #model.model_2 = model.model_2.load(name='model_2', folder=dir_tmp,)
+        model.model_5 = model.model_5.load(name='model_5', folder=dir_tmp,)
         shutil.rmtree(dir_tmp)
         logger.info('Load AutoML')
         return(model)