+++ automl_alex/models/model_catboost.py
@@ -2,6 +2,7 @@ from automl_alex.base import ModelBase
 import catboost
 from catboost import Pool
 import numpy as np
+import pandas as pd
 
 
 class CatBoost(ModelBase):